Removing factories and added validation that the estate was bought on assembling for bd

+++ HouseMatch/src/main/java/ca/ulaval/glo4003/b6/housematch/domain/estate/Estate.java
@@ -115,7 +115,7 @@ public class Estate {
    }
 
    public void editPrice(Integer newPrice) {
-      if (this.price != newPrice) {
+      if (this.price.intValue() != newPrice.intValue()) {
          this.priceHistory.add(this.price);
          this.price = newPrice;
       }
